fix data generation
